feat: do it with the admin cli instead
