+++ Applications/AdminCli/src/AdminCli/Commands/Database/Types/RelationshipTemplateExport.cs
@@ -1,0 +1,10 @@
+﻿namespace Backbone.AdminCli.Commands.Database.Types;
+
+public class RelationshipTemplateExport
+{
+    public required string TemplateId { get; set; }
+    public required string CreatedBy { get; set; }
+    public string? CreatedByClientName { get; set; }
+    public required DateTime CreatedAt { get; set; }
+    public required DateTime? AllocatedAt { get; set; }
+}
