+++ Applications/AdminCli/src/AdminCli/Commands/RootCommand.cs
@@ -1,5 +1,6 @@
 using Backbone.AdminCli.Commands.Announcements;
 using Backbone.AdminCli.Commands.Clients;
+using Backbone.AdminCli.Commands.Database;
 using Backbone.AdminCli.Commands.Tiers;
 using Backbone.AdminCli.Commands.Tokens;
 
@@ -7,9 +8,10 @@ namespace Backbone.AdminCli.Commands;
 
 public class RootCommand : System.CommandLine.RootCommand
 {
-    public RootCommand(ClientCommand clientCommand, TierCommand tierCommand, AnnouncementCommand announcementCommand, TokenCommand tokenCommand)
+    public RootCommand(ClientCommand clientCommand, TierCommand tierCommand, AnnouncementCommand announcementCommand, TokenCommand tokenCommand, DatabaseCommand databaseCommand)
     {
         AddCommand(clientCommand);
+        AddCommand(databaseCommand);
         AddCommand(tierCommand);
         AddCommand(announcementCommand);
         AddCommand(tokenCommand);
