+++ Applications/AdminCli/src/AdminCli/Program.cs
@@ -5,10 +5,15 @@ using Backbone.AdminCli.Configuration;
 using Backbone.BuildingBlocks.API.Extensions;
 using Backbone.BuildingBlocks.Application.QuotaCheck;
 using Backbone.Modules.Announcements.Module;
+using Backbone.Modules.Challenges.Module;
 using Backbone.Modules.Devices.Domain.Entities.Identities;
 using Backbone.Modules.Devices.Infrastructure.OpenIddict;
 using Backbone.Modules.Devices.Infrastructure.Persistence.Database;
 using Backbone.Modules.Devices.Module;
+using Backbone.Modules.Files.Module;
+using Backbone.Modules.Messages.Module;
+using Backbone.Modules.Relationships.Module;
+using Backbone.Modules.Synchronization.Module;
 using Backbone.Modules.Tokens.Application;
 using Backbone.Modules.Tokens.Module;
 using Microsoft.AspNetCore.Identity;
@@ -71,7 +76,12 @@ public class Program
 
         services
             .AddModule<AnnouncementsModule, Modules.Announcements.Application.ApplicationConfiguration, Modules.Announcements.Infrastructure.InfrastructureConfiguration>(configuration)
+            .AddModule<ChallengesModule, Modules.Challenges.Application.ApplicationConfiguration, Modules.Challenges.Infrastructure.InfrastructureConfiguration>(configuration)
             .AddModule<DevicesModule, Modules.Devices.Application.ApplicationConfiguration, InfrastructureConfiguration>(configuration)
+            .AddModule<FilesModule, Modules.Files.Application.ApplicationConfiguration, Modules.Files.Infrastructure.InfrastructureConfiguration>(configuration)
+            .AddModule<MessagesModule, Modules.Messages.Application.ApplicationConfiguration, Modules.Messages.Infrastructure.InfrastructureConfiguration>(configuration)
+            .AddModule<RelationshipsModule, Modules.Relationships.Application.ApplicationConfiguration, Modules.Relationships.Infrastructure.InfrastructureConfiguration>(configuration)
+            .AddModule<SynchronizationModule, Modules.Synchronization.Application.ApplicationConfiguration, Modules.Synchronization.Infrastructure.InfrastructureConfiguration>(configuration)
             .AddModule<TokensModule, ApplicationConfiguration, Modules.Tokens.Infrastructure.InfrastructureConfiguration>(configuration);
 
         var containerBuilder = new ContainerBuilder();
